test: introduce stress test for memtx MVCC DDL
DDL with memtx mvcc enabled used to crash a lot until the previous
commits have fixed it. To make sure it's stable now, the commit
introduces a stress test that executes various DDL and DML operations
concurrently. The test doesn't check serialization of transactions, the
only goal is to make sure that Tarantool does not crash.

Along the way, the commit introduces a new error injection that disables
yield while building an index. The problem is any concurrent DML with
DDL building an index will abort it if MVCC is enabled, so the error
injection is needed to make index build successful during the stress
test.

Follow-up #10146

NO_CHANGELOG=test
NO_DOC=test