applier: fix memory leak when replica is removed from cluster by master
When replica is deleted from cluster it receives the deletion statement
and then connection to master is closed by master. Since the commit
25aac111ce10 ("box: apply deletion of replica from `_cluster` on the
deleted replica") we kill applier to master on the deletion. Thus the
exit message produced due to EOF on connection to master is never
executed. This message in next in queue after the deletion statement.
As a result the exit message diag with SocketError from applier thread
is not freed.

Closes #10490
Part-of #10211

NO_TEST=covered by existing tests
NO_DOC=bugfix