box: make asynchronous transactions bump the limbo's confirmed LSN
Currently, we cannot detect a split-brain situation on the old leader when
a promote arrives, and the old leader has committed asynchronous
transactions that are not confirmed by the promote. We can only detect this
on the new leader if he sees asynchronous transactions replicated to him
from the old leader.

This is because asynchronous transactions do not bump the confirmation
boundary of the synchronous queue, which is used to detect such a
split-brain situation. Even though asynchronous transactions can completely
bypass the synchronous queue, they must still bump the confirmation
boundaries as if they were instantly confirmed (since asynchronous
transactions are per se synchronous transactions with quorum 1).

There are several cases when we must not bump the confirmation boundary:
* when limbo does not have an owner — this would impose a performance
penalty on purely asynchronous workloads;
* when the instance is anonymous — it is read-only;
* when the transaction is fully local not — it is not replicated and
is associated with the local vclock component;
* when the transaction is fully nop — it comes from an old term, is rolled
back on lagging instances, and nopped on up-to-date instances for vclock c
onsistency;
* when the transaction is a snapshot row — during recovery, we have no
information about the rows' original LSNs, so all snapshot rows have the
snapshot's signature as their LSN, i.e., are not related to any vclock
component — however, the latest confirmed LSN is written to the snapshot as
the promote request, so we do not lose the confirmation boundary.

We will now need to wait for all the preceding transactions to receive an
LSN when checking the confirmation boundary of the limbo. This condition is
stronger than waiting for all the limbo entries to receive an LSN, so we
can safely drop the related code. Since now we will also have 2 points from
which `wal_sync` is called, we need to adjust the corresponding error
injection usage and introduce a new countdown injection.

Closes #10528
Needed for #10459

NO_DOC=<covered by existing documentation>