box: refactor limbo LSN computation from journal write callback
In the scope of #10528 we are going to add additional logic around the
limbo LSN computation in the journal write callback. The computation is a
complicated snippet of code on its own, so let's refactored it into a
separate helper.

Needed for #10528

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>