vinyl: fix ERRINJ_VY_DELAY_PK_LOOKUP
Enabling `ERRINJ_VY_DELAY_PK_LOOKUP` makes Vinyl yield in a place where
it wouldn't normally do. If the transaction is aborted in the meantime,
we'll get the assertion failure:

```
./src/box/vy_point_lookup.c:219: vy_point_lookup: Assertion 'tx == NULL || tx->state == VINYL_TX_READY' failed.
```

To prevent this from happening, let's replace this invalid error
injection with the new one `ERRINJ_VY_POINT_LOOKUP_DELAY` that injects
a delay to `vy_point_lookup()` before reading disk. This doesn't have
exactly the same effect as the old error injection because it also
delays direct lookups in the primary index. Fortunately, the old error
injection is used in the only test, where the new one works as expected
if we make the secondary index created in the test non-unique and enable
deferred writes (this makes the `s:replace{2, 2}` statement bypass
a lookup in the primary index).

Also, let's replace `VY_POINT_ITER_WAIT` with the new error injection
because they have very a similar meaning and `VY_POINT_LOOKUP_DELAY`
works in the test using it with a very small adjustment (we need to
clear it explicitly after `box.snapshot()`).

Closes #10517

NO_DOC=errinj fix
NO_CHANGELOG=errinj fix