test: specify the dir in the cluster helper
The cluster helper in `test/config-luatest/cluster.lua` is used to
define the clusters in an easy way. Though every time the `cluster.new`
is called it creates a new temporary directory, which is not good when
you're writing a test setting up the directory before the cluster
start.

This patch adds an `opts` object as the fourth parameter for
`cluster.new` and makes the function use `opts.dir` as a key for
specifying the directory to start the cluster.

Example:
```
    local dir = '/some/long/dir'
    local cluster_config = cbuilder:new(remote_config)
        :use_replicaset('r-001')
        :add_instance('i-001', {})
        :config()

    -- The cluster will be started in a new temp directory.
    local cluster = cluster.new(g, cluster_config)

    -- The cluster will be started in '/some/long/dir'.
    local cluster = cluster.new(g, cluster_config, nil, { dir = dir })
```

Needed for testing tarantool/tarantool-ee#870.

NO_CHANGELOG=internal testing helper change
NO_TEST=internal testing helper change
NO_DOC=internal testing helper change