vinyl: use VERBOSE level for logging ranges
Whenever a range is compacted, split, or coalesced, we log the range
boundaries. This gets really annoying if there's an index that has
a lot of key parts or contains binary strings. Let's lower the level
used for logging these events down to VERBOSE so that they are not
shown by default but can be enabled if needed.

Closes #10524

NO_DOC=bug fix