vinyl: fix suboptimal msgpack assert in upsert
Upsert operation normalizes its operations - subtracts the index
base from the fieldnos. If the fieldno is negative, it is simply
encoded back. If positive - the subtraction happens and then
encoding back.

The problem was that the code assumed MP_INT is always negative,
and would blindly encode it without changes. The assumption wasn't
right - MP_INT can be positive. Need to handle that.

Part of #9965

NO_DOC=bugfix
NO_CHANGELOG=later