txn: set `TXN_FORCE_ASYNC` flag for nopped out synchronous transactions
We nop out incoming synchronous transactions from old terms or conflicting
transactions in `replication_skip_conflict`, and they always bypass the
synchronous queue.

In the scope of #10528 such transactions should also not bump the confirmed
boundary of the synchronous queue, when committed. Figuring out that a
transaction is fully nop requires traversing it's list of statement, and we
already do that during the journal entry creation, and they are logically
equivalent to `TXN_FORCE_ASYNC` transactions — so let's set that flag for
them too.

Needed for #10528

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>