datetime: fix suboptimal msgpack decoding
Datetime interval object, when being decoded from msgpack, didn't
consider MP_INT allowing positive values. In one places, where a
number of any sign in int32_t was allowed, the MP_INT decoded
value was only checked for INT32_MIN, assuming that max is 0. But
in fact it had to be checked for INT32_MAX too, as it can be > 0.

Part of #9965

NO_DOC=bugfix
NO_CHANGELOG=later