perf: have explicit main() in tuple bench
It is a step towards removing the global singletons completely.
They make it super hard to customize any bench, because they all
have to use the same globals right now. It would make problematic
to introduce a bench for tuple hashing in the next commits.

In scope of #9965

NO_DOC=internal
NO_CHANGELOG=internal
NO_TEST=it is a test