perf: drop singletons from tuple bench
They made it super hard to customize any bench, because all
benches had to use the same globals. It wasn't easy, for example,
specify a new kind of key def with a different set of fields. All
the code done by those globals would have to be duplicated. But
now they are more generic and reusable.

In scope of #9965

NO_DOC=internal
NO_CHANGELOG=internal
NO_TEST=it is a test