perf: fix wrong key_part_def initialization
key_part_def was initialized with zeros, which is wrong. It makes
it have ON_CONFLICT_ACTION_NONE set (because it is zero), which
isn't default. This led to some fields being nullable
unintentionally.

Better use the proper key_part_def_default to initialize those
defs.

In scope of #9965

NO_DOC=internal
NO_CHANGELOG=internal
NO_TEST=it is a test