test: use built tarantool instead a system one
The test gh_5747_crash_multiple_args_test.lua uses a `tarantool`
binary that is available in a current PATH environment variable.
In some cases, PATH may point to a system `tarantool` executable
and this could be unexpected for those who run the test. The patch
set a path to `tarantool` executable derived from a command-line
used to run the test.

NO_CHANGELOG=codehealth
NO_DOC=codehealth