cmake: shuffle testcases in luatest's tests
luatest has an option `--shuffle` that allows to reorder testcases
before execution. This can be useful to detect a test that passes
just because it happens to run after an unrelated test that leaves
the system in a favourable state. Patch enables tests shuffling
with predefined random seed generated by CMake [1] and a fixes for
a tests were failed after enabling shuffle mode. Seed is useful
for reproducing a problems related to the specific test order.

1. https://cmake.org/cmake/help/latest/command/string.html#random

The following tests FAILED:
901 - test/app-luatest/http_client_http_version_test.lua (Failed)
1101 - test/box-luatest/gh_8122_excess_conflict_secondary_test.lua (Failed)
1104 - test/box-luatest/gh_8147_tuple_formats_in_iproto_responses_test.lua (Failed)
1116 - test/box-luatest/gh_8633_tuple_formats_in_iproto_requests_test.lua (Failed)
1355 - test/replication-luatest/gh_8931_box_info_election_leader_name_test.lua (Failed)
1703 - test/sql-luatest/interval_test.lua (Failed)

NO_CHANGELOG=testing
NO_DOC=testing
NO_TEST=testing