memtx: only store in-index tuples in the cleaner
Currently the snapshot cleaner works this way:
1. It saves the tuples clarified by all existing space stories in a
   hash table associating story tuples with clarified ones. So we can
   get the tuple visible to a transaction by looking up an in-index one
   in the hash table.
2. We also have space alter statements introduced in the commit
   a532e375fe3703a210a58e1cc4b57fdc608d34c2 ("box: exclude uncommitted
   alter records from snapshot"), which are also handled in the table.

The cleaner itself is only used with in-index tuples, so we could only
traverse top PK stories to fill it. But because of the point 2 we have
to associate historical tuples too since the alter statement always
associates a current tuple with a historical one, so we have to look
it up in the cleaner hash table too to get the actual clear tuple.

Let's only handle alter statements in no-MVCC mode, this way we can
also stop storing associations from historical tuples in the cleaner
and only associate in-index tuples in it.

Needed for tarantool/tarantool-ee#858

NO_DOC=internal
NO_TEST=covered by existig tests
NO_CHANGELOG=internal