memtx: itroduce `exhaust_after` and `step_count` iterator members
The EQ/REQ tree iterators in the memtx read view perform comparison
of the search key with each next tuple using the `key_data` member.
The thing is, since a read view of a tree is immutable, we can just
perform the second lookup to find the last tuple and only compare
tuple pointers on each step. This will reduce the total overhead of
iteration over tuples with huge field count, because the equal key
comparison is expensive for them in read views now since we can't
use the tuple format.

The first step towards this is introducing a new `exhaust_after`
field to be used instead of the `key_data` in case of partial key
search, and `step_count` for heuristics on whether to optimize the
iteration or not.

TODO: caculation of the optimization threshold.

Part of tarantool/tarantool-ee#921

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal