txn: ack limbo owner WAL writes to the limbo asynchronously
Currently, a synchronous transaction must always wait for its WAL write to
finish, acknowledge itself to the synchronous queue, and potentially wait
for a synchronous CONFIRM request WAL write.

In the scope of #10583, we want to support asynchronous wait modes for
synchronous transactions. To enable the none mode, we need to acknowledge
our own WAL writes for the synchronous queue asynchronously by:
1. Moving the acknowledgment to the callback for finishing transaction WAL
   write;
2. Detaching the CONFIRM request WAL writing from the acknowledgment, and
   making it asynchronous.

To enable the latter, we introduce a worker fiber for the limbo, which is
responsible for closing the gap between the persisted confirmed LSN and the
volatile confirmed LSN by trying to write a CONFIRM request to the WAL, and
retrying it on failure.

To be able to test the limbo worker behaviour in the event of a concurrent
PROMOTE/DEMOTE request processing, we add a new
`ERRINJ_TXN_LIMBO_WORKER_DELAY` error injection.

Closes #10628

NO_CHANGELOG=<no behaviour change>
NO_DOC=<no behaviour change>