limbo: refactor synchro request writing without panicking out
In the scope of #10628 we will need to be able to retry writing failed
CONFIRM requests. Currently, failed synchronous replication requests
unconditionally cause panic, so let's refactor the logic of writing a
synchronous replication request without panicking into a separate helper.

The patch is cherry-picked to 3.2 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790. The original commit also didn't compile. The
cherry-picked version doesn't change it and applies the commit as
is.

Needed for #10628
Needed for #11790

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>

(cherry picked from commit f7650b54dfd020a50a37fcfcd033e5904b7c1c15)