upgrade: allow to downgrade from 3.0.0 with names set
Currently it's impossible to downgrade without removing
instance/replicaset/cluster names, which is hard to do, since it
requires reconfiguring all instances with force_recovery.

However, it's not the user's fault, that the names were set, config
module sets them automatically as soon as it can, which interferes with
downgrade.

We must make the downgrade process from 3.0 as easy as we can.
In order to achieve this from now on it's allowed to downgrade, while
the names are still set, we don't require user to drop them, we also
don't drop them ourselves, all names remain untouched.

In the following commits, it'll be also tested, that the names properly
works, while the schema is < 3.0.0. We'll intentionally allow names
before upgrade in order to simplify the work for tarantool modules,
which won't have to deal with dual identification: uuid and names.

Closes #10150

NO_DOC=bugfix