gc: fix gc consumer erasure on old schema
If _gc_consumers space doesn't exist, then deletion of
the replica from the _cluster causes segmentation fault,
when we try to delete row before this instance manages
to connect and create consumer.

Let's fix that by not trying to do anything with consumer on its
erasure if consumer doesn't exist and if schema doesn't support
persistent gc.

It's needed, since in the next commit it will be allowed
to make DDL before upgrade and names will on the schema below 3.0.0.

Needed for #10520

Co-authored-by: Andrey Saranchin <Andrey22102001@gmail.com>

NO_DOC=bugfix
NO_CHANGELOG=<cannot be reproduced without internal functions>