net_box: fix a crash when a trigger deletes itself
In 6d88274 we rewrote Lua module `internal.trigger` that is used in
`net.box` and possibly some external modules. Old implementation held
all triggers in a Lua table, so deleting a trigger on the fly wasn't a
problem. Now we store all triggers in a list and simply iterate over it
when the triggers are fired, so the trigger list became non-resistent to
modifications on the fly. In order to fix it, let's simply use
`trigger_run` - it is resistant to the list modifications.

Closes #10622

NO_DOC=bugfix

(cherry picked from commit 78cfc5ef4cf43bf9d53dc51d0f843bee350609fd)