memtx: record getting the index size in the MVCC
Currently getting the index size (aka calling the index.len() method)
is not recorded in the MVCC. This has a side effect that if the size
if an index is changed after the transaction first inspected it, it's
not conflicted and may continue performing writing operations on the
space. That's a problem since in this case the transaction can perform
a set of operations assuming that at the time it commits the size of
the index is the same as been checked, which may not be true.

Let's fix this by informing the MVCC system about the index size check
performed by a transaction.

Closes #10149

NO_DOC=bugfix