upgrade: fix downgrading to 2.11.5 schema version
Some time ago we decided to introduce a new schema version to the 2.11
release series, 2.11.5. It doesn't carry any changes compared to the
previous version, 2.11.1, and exists solely to mark the corresponding
Tarantool version as supporting persistent names.

This happened in commit 46cac24cd569 ("upgrade: introduce 2.11.5 schema
version"), however that commit was only pushed to the 2.11 branch, so
the 3.x series still don't know anything about the 2.11.5 schema
version and fail to downgrade to it. Let's fix this.

Closes #11204

NO_DOC=bugfix