box/lua: introduce space:insert_arrow(arrow_record_batch)
The new Lua type arrow_record_batch is introduced. This is a simple
wrapper over an ArrowArray with a corresponding ArrowSchema. Currently
it is possible to create it only from the MsgPack extension MP_ARROW.

Also the new space method `insert_arrow()` is introduced. It takes an
arrow record batch as an input and calls `box_insert_arrow()`.

Follow-up #10508
Needed for tarantool/tarantool-ee#865

NO_DOC=No sense in mentioning in CE
NO_CHANGELOG=No sense in mentioning in CE