upgrade: fix upgrading from schema 1.6.8
This commit fixes upgrading schema from 1.6.8:

1. Fix updating empty password for users. In 1.6 credentials were array
   in _user, in 1.7.5 they became map.

2. Start to automatically change format of user spaces: previously it
   was {'a', type = 'number'}, now it's {name = 'a', type = 'number'}.
   Make sure, that space format is encoded as array.

This commit also introduces the test, which checks, that we can
properly upgrade from 1.6.8 to the latest versions, at least in basic
cases.

There're still a lot of side cases, which are not covered by the
upgrade triggers: e.g. user can create space with format {1, 2, 3} on
schema version 1.6, but we don't have enough info to figure out, which
format should it be and we'll fail to start the instance. User has
to have properly configured schema on 1.6.8 in order to upgrade.

Closes #10180

NO_DOC=bugfix