tuple: fix bad hashing of suboptimal int msgpack
Tuple and key hashing in tuple_hash.cc could return different
hashes for the same integer encoded in different ways.

In MessagePack it is possible to encode one integer in up to 9
ways, all being valid.

Tarantool must either prohibit suboptimal integer encoding, or
properly support it, but not ignore like it was before.

This commit chose the support way. Tuple hashing code now repacks
the integers in the optimal way before taking their hash. This is
enough for memtx.

For Vinyl the situation was a bit different - it stores hashes on
disk, for the bloom filters. In order for the filters generated by
the older versions of Tarantool to be usable, there is still
support of incorrect hashing functions. Over time, after some
compactions, the old filters are going to be washed away and
replaced by the new ones automatically.

The new hash functions not only aren't slower, but are even
slightly faster. The reason is that the hashing algorithm seems to
have a heavy penalty for all buffers whose size isn't a power of
2. The optimally encoded MessagePack integers usually take an
uneven number of bytes: 1, 2, 3, 5, 9. Only numbers in the ranges
[-128, -32) and (127, 255) take a buffer of an even size.

The new hashing functions use the same hash algorithm, but
re-encode the MesssagePack integers without their headers. Hence
operating always on the buffers of sizes 1, 2, 4, 8 - smaller and
only one of them is uneven.

The speed up isn't large though. It can make visible effect on the
indexes consisting of more than one integer key part (single
unsigned key part uses a totally different algorithm, which wasn't
changed). The observable speed increase is reproducible in
tuple.perftest with `--benchmark_filter="tuple_tuple_hash"`
argument, and was seen to be up to 11% on just tuple_hash()
function alone. In a realistic case the increase might be much
less visible when full overhead of a complete request is
considered.

Closes #9965

NO_DOC=bugfix