engine: fix sending raft state during META_JOIN
Engine (and previously relay) didn't send raft state during META_JOIN
stage, as it was marked with GROUP_LOCAL id and skipped by the stream
function. This may cause split brain as replica doesn't have any info
about term before the time, it gets it in the beginning of the
SUBSCRIBE (see sent_raft_term).

In order not to repeat the same error in the future let's automatically
decode/encode state with the correct group id.

Closes #10089

NO_DOC=bugfix