box: introduce box.ctl.replica_gc
The commit introduces a new function `box.ctl.replica_gc`. It can be
used to cleanup state of a replica that is gone for a while: it frees
replica's WAL GC state if it is disconnected and drops the replica if
it is anonymous. If the replica is connected, an error is thrown.

In order to call `gc_erase_consumer` under the hood, the commit
populates it with autocommit mode so that it can be called without an
active transaction. However, underlying `txn_current_stmt` call is not
appropriate in this case - current statement is actually one that is
being executed: for example, we used to call this function in
`_cluster:on_replace` - in this case current statement is the replace in
`_cluster`. Since we don't have current statement in autocommit mode,
let's use last statement (the actual deletion in `_gc_consumers`) instead.

Part of #10755

NO_CHANGELOG=later
NO_DOC=later