config: fix scaling a replicaset down to one node
Before this patch there was a mistake in the YAML configuration
processing, when an instance removal is not taken into account in
generation of a list of upstreams. It occurs when the updated YAML
configuration has only one instance in the given replicaset.

Fixes #10716

NO_DOC=the new behavior is an expected one, the change is a bug fix

(cherry picked from commit 1a85b917c32ec5cc873f03cf8fb4f2b466bbb54a)