replication: store checkpoint reference in replica
On checkpoint join, we reference snapshot right before the join and
dereference it right after. The point of checkpoint join is that you can
try to fetch it again from the same snapshot on failure, so it makes
sense to dereference it only when the replica has completed the join or
has gone - that's what the commit does.

Actually, the commit moves checkpoint referencing logic from checkpoint
join internals (implemented in Tarantool EE) because we need to obtain
vclock of the checkpoint before registering replica's WAL GC consumer.
After the change, `prepare_checkpoint_join` and
`complete_checkpoint_join` functions are not needed anymore so theirs
stubs are removed.

Part of tarantool/tarantool-ee#742

NO_TEST=in EE
NO_CHANGELOG=later
NO_DOC=later