box: allow to access to the granted role in _vuser space
Prior to this patch it was not possible for a user to see any role in
the `_vuser` space except the one created by him and the `public` one.

Let's also make it possible to see roles that he is granted with and
all the roles that role is granted transitively. Now calls to the
`box.schema.user.info()` functions won't fail if a user has a non-
'public' role granted.

Closes #10057

NO_DOC=bugfix

(cherry picked from commit 0280afce5cfaf46d4383f35613de683578da4e8a)