box: allow to access to the granted role in _vuser space
Prior to this patch it was not possible for a user to see any role in
the `_vuser` space except the one created by him and the `public` one.

Let's also make it possible to see roles that he is granted with, so
calls to the `box.schema.user.info()` functions won't fail if a user
has a non-'public' role granted.

Closes #10057

NO_DOC=bugfix