config: fix handling of audit_log.spaces removal
If the option is removed from the YAML configuration, `config:reload()`
should add an alert that a non-dynamic option is changed and it will be
applied after a restart. This patch works the problem around. The root
of the problem will be fixed in #10756.

Fixes tarantool/tarantool-ee#964

NO_DOC=bugfix